+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 7.0.14-SNAPSHOT / WIP
 ==================
+Bug fixes:
+* [OLMIS-8042](https://openlmis.atlassian.net/browse/OLMIS-8042): Total received quantity column is readonly when requisition is from stock cards. For every program other than TB Monthly
 
 7.0.13 / 2024-10-31
 ==================